+++ packages/aws-lambda/serverless.js
@@ -39,7 +39,7 @@ const defaults = {
   handler: "handler.hello",
   runtime: "nodejs10.x",
   env: {},
-  region: "us-west-2"
+  region: "us-east-1"
 };
 
 class AwsLambda extends Component {