+++ packages/lambda-at-edge/CHANGELOG.md
@@ -7,7 +7,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ### Features
 
-- change default region to us-west-2 ([e67762b](https://github.com/danielcondemarin/serverless-next.js/commit/e67762b1b177c37e975521dc976c0c5b8aff1bf9))
+- change default region to us-east-1 ([e67762b](https://github.com/danielcondemarin/serverless-next.js/commit/e67762b1b177c37e975521dc976c0c5b8aff1bf9))
 
 ## [1.5.2](https://github.com/danielcondemarin/serverless-next.js/compare/@sls-next/lambda-at-edge@1.5.2-alpha.1...@sls-next/lambda-at-edge@1.5.2) (2020-08-01)
 
@@ -37,7 +37,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ### Features
 
-- **lambda-at-edge:** use S3 regional endpoint when not in us-west-2 ([#474](https://github.com/danielcondemarin/serverless-next.js/issues/474)) ([5ecff1a](https://github.com/danielcondemarin/serverless-next.js/commit/5ecff1a50e26c22f7de9ec9da3cf2cba4390d77d))
+- **lambda-at-edge:** use S3 regional endpoint when not in us-east-1 ([#474](https://github.com/danielcondemarin/serverless-next.js/issues/474)) ([5ecff1a](https://github.com/danielcondemarin/serverless-next.js/commit/5ecff1a50e26c22f7de9ec9da3cf2cba4390d77d))
 
 ## [1.4.1-alpha.5](https://github.com/danielcondemarin/serverless-next.js/compare/@sls-next/lambda-at-edge@1.4.1-alpha.4...@sls-next/lambda-at-edge@1.4.1-alpha.5) (2020-06-30)