+++ packages/serverless-component/CHANGELOG.md
@@ -7,7 +7,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ### Features
 
-- change default region to us-west-2 ([e67762b](https://github.com/danielcondemarin/serverless-next.js/commit/e67762b1b177c37e975521dc976c0c5b8aff1bf9))
+- change default region to us-east-1 ([e67762b](https://github.com/danielcondemarin/serverless-next.js/commit/e67762b1b177c37e975521dc976c0c5b8aff1bf9))
 
 # [1.16.0](https://github.com/danielcondemarin/serverless-next.js/compare/@sls-next/serverless-component@1.16.0-alpha.1...@sls-next/serverless-component@1.16.0) (2020-08-01)
 
@@ -32,7 +32,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 ### Bug Fixes
 
 - **serverless-component:** fix "main" package.json property ([96d510d](https://github.com/danielcondemarin/serverless-next.js/commit/96d510d3e21f6fe84c02425f2c8c6f311b90665f))
-- **serverless-component:** fixes 307 errors when using bucket region outside us-west-2 ([#495](https://github.com/danielcondemarin/serverless-next.js/issues/495)) ([561e05f](https://github.com/danielcondemarin/serverless-next.js/commit/561e05f9ba9763fab4185b34adcde61dfcbb0a21))
+- **serverless-component:** fixes 307 errors when using bucket region outside us-east-1 ([#495](https://github.com/danielcondemarin/serverless-next.js/issues/495)) ([561e05f](https://github.com/danielcondemarin/serverless-next.js/commit/561e05f9ba9763fab4185b34adcde61dfcbb0a21))
 
 ## [1.15.1](https://github.com/danielcondemarin/serverless-next.js/compare/@sls-next/serverless-component@1.15.0...@sls-next/serverless-component@1.15.1) (2020-07-11)